Reduce number of CLI end-to-end testing, factor out as test_cli_import.
Currently CSV all use-case has end-to-end testing, which is too much.
